Integrate globalbrain-node module for calling GlobalBrain.jl from javascript
