+++ app/root.tsx
@@ -1,6 +1,5 @@
 import { useForm } from '@conform-to/react'
 import { parse } from '@conform-to/zod'
-import { cssBundleHref } from '@remix-run/css-bundle'
 import {
 	type DataFunctionArgs,
 	type HeadersFunction,
@@ -56,6 +55,9 @@ import { getTheme, setTheme, type Theme } from './utils/theme.server.ts'
 import { makeTimings, time } from './utils/timing.server.ts'
 import { getToast } from './utils/toast.server.ts'
 import { useOptionalUser, useUser } from './utils/user.ts'
+// import { cssBundleHref } from '@remix-run/css-bundle'
+
+const cssBundleHref = null
 
 export const links: LinksFunction = () => {
 	return [
