+++ app/utils/env.server.ts
@@ -4,6 +4,7 @@ const schema = z.object({
 	NODE_ENV: z.enum(['production', 'development', 'test'] as const),
 	APP_DATABASE_PATH: z.string(),
 	APP_DATABASE_URL: z.string(),
+	GB_DATABASE_PATH: z.string(),
 	SESSION_SECRET: z.string(),
 	INTERNAL_COMMAND_TOKEN: z.string(),
 	HONEYPOT_SECRET: z.string(),
