+++ package.json
@@ -26,7 +26,8 @@
     "test:e2e:run": "cross-env CI=true playwright test",
     "test:e2e:install": "npx playwright install --with-deps chromium",
     "typecheck": "tsc",
-    "validate": "run-p \"test -- --run\" lint typecheck test:e2e:run"
+    "validate": "run-p \"test -- --run\" lint typecheck test:e2e:run",
+    "postinstall": "remix-esbuild-override"
   },
   "eslintIgnore": [
     "/node_modules",
@@ -117,6 +118,7 @@
     "@remix-run/serve": "^2.1.0",
     "@remix-run/testing": "^2.1.0",
     "@sly-cli/sly": "^1.4.10",
+    "@socialprotocols/globalbrain-node": "file:../GlobalBrain.jl/globalbrain-node",
     "@testing-library/jest-dom": "^6.1.3",
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.5.1",
@@ -156,6 +158,7 @@
     "prettier": "3.2.5",
     "prettier-plugin-sql": "^0.15.1",
     "prettier-plugin-tailwindcss": "^0.5.4",
+    "remix-esbuild-override": "^3.1.0",
     "remix-flat-routes": "^0.5.11",
     "source-map-support": "^0.5.21",
     "tsx": "^3.14.0",
