+++ remix.config.js
@@ -1,4 +1,13 @@
+import { withEsbuildOverride } from 'remix-esbuild-override'
 import { flatRoutes } from 'remix-flat-routes'
+
+// https://github.com/remix-run/remix/issues/1823
+withEsbuildOverride(option => {
+	if (!option.external && option.bundle) {
+		option.external = ['nock', 'mock-aws-s3', 'node-pre-gyp', 'node-gyp']
+	}
+	return option
+})
 
 /**
  * @type {import('@remix-run/dev').AppConfig}
